samples/agent/adk/mcp_app_proxy/pyproject.toml
# Copyright 2026 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "mcp_app_proxy"
version = "0.1.0"
description = "A reference agent that proxies an MCP app"
input-modes = ["text"]
output-modes = ["text"]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "litellm",
    "jsonschema>=4.0.0",
    "a2ui-agent",
    "uvicorn",
    "starlette",
    "mcp",
    "anyio"
]

[project.scripts]
mcp_app_proxy = "mcp_app_proxy:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
a2ui-agent = { path = "../../../../agent_sdks/python", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["."]
